# PERF: optimize conversion from boolean Arrow array to masked BooleanArray