--- a/pandas/tests/arrays/masked/test_arrow_compat.py
+++ b/pandas/tests/arrays/masked/test_arrow_compat.py
@@ -55,12 +55,19 @@ def test_arrow_from_arrow_uint():
 
 
 @td.skip_if_no("pyarrow", min_version="0.16.0")
-def test_arrow_sliced():
+def test_arrow_sliced(data):
     # https://github.com/pandas-dev/pandas/issues/38525
     import pyarrow as pa
 
-    df = pd.DataFrame({"a": pd.array([0, None, 2, 3, None], dtype="Int64")})
+    df = pd.DataFrame({"a": data})
     table = pa.table(df)
     result = table.slice(2, None).to_pandas()
     expected = df.iloc[2:].reset_index(drop=True)
+    tm.assert_frame_equal(result, expected)
+
+    # no missing values
+    df2 = df.fillna(data[0])
+    table = pa.table(df2)
+    result = table.slice(2, None).to_pandas()
+    expected = df2.iloc[2:].reset_index(drop=True)
     tm.assert_frame_equal(result, expected)