@jorisvandenbossche 2021-04-20
> i'm getting significant different timings and showing a much bigger improvement (30x with missing values) 

Whoops, I was using my debug build of arrow, which makes things generally much slower, that should explain it. 
But so all the better that the speed-up is larger! ;)